refacto: create filter2d_api
